Add conversation search overlay and fix post-merge dep upgrades
Adds a search icon next to the sidebar toggle (macOS floating + Linux/Windows
titlebar) that opens an overlay dialog with a focused input and a live-filtered,
most-recent-first list of all conversations. Selecting a result reuses the
existing SidebarEvent::SelectConversation flow so the conversation opens in the
main view and the sidebar marks it active.

Also unblocks the build after merging main by applying a batch of pre-existing
dependency-upgrade migrations:

- ropey: pin to =2.0.0-beta.1 to match gpui-component's required version
- sha2 0.11: replace `format!("{:x}", hasher.finalize())` with `hex::encode(...)`
  in math/mermaid renderers, pdf_thumbnail, pdf_to_image_tool, auto_updater
- pdfium-render 0.9: migrate `page.get(_ as u16)` to `as i32`, handle
  `bitmap.as_image()` now returning Result, and teach `create_pdfium()` to
  reuse the process-global singleton when bindings are already initialized
  (fixes concurrent test re-init panic)
- bollard: adapt exposed_ports/CreateImageOptions to the new query-parameter
  shape in the Docker sandbox

+++ crates/chatty-core/src/tools/pdf_info_tool.rs
@@ -152,7 +152,7 @@ fn get_pdf_info(pdf_path: &std::path::Path) -> Result<PdfInfoResult, ToolError> 
     let mut pages = Vec::new();
 
     for i in 0..page_count.min(MAX_INFO_PAGES) {
-        if let Ok(page) = document.pages().get(i as u16) {
+        if let Ok(page) = document.pages().get(i as i32) {
             let width_pt = page.width().value;
             let height_pt = page.height().value;
             pages.push(PageInfo {
